requirements.txt
torch>=1.9.0
transformers>=4.10.0
streamlit>=1.0.0
pandas>=1.3.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
seaborn>=0.11.0
python-dotenv>=0.19.0
setuptools>=58.0.0
pytest>=6.2.5